+++ python/marvin/web/lib/src/heatmap.js
@@ -10,7 +10,7 @@
 class HeatMap {
 
     // Constructor
-    constructor(mapdiv, data, title, galthis) {
+    constructor(mapdiv, data, title, plotparams, galthis) {
         if (data === undefined) {
             console.error('Must specify input map data to initialize a HeatMap!');
         } else if (mapdiv === undefined) {
@@ -19,6 +19,7 @@ class HeatMap {
             this.mapdiv = mapdiv; // div element for map
             this.data = data; // map data
             this.title = title; // map title
+            this.plotparams = plotparams; // default plotting parameters
             this.galthis = galthis; //the self of the Galaxy class
             this.parseTitle();
             this.initMap();
@@ -29,7 +30,7 @@ class HeatMap {
 
     // test print
     print() {
-        console.log('We are now printing heatmap for ', this.title);
+        console.log('We are now printing heatmap for ', this.title, 'test');
     };
 
     // Parse the heatmap title into category, parameter, channel
@@ -76,41 +77,34 @@ class HeatMap {
                 var val = values[ii][jj];
 
                 if (mask !== null) {
-                    var noValue = (mask[ii][jj] & Math.pow(2, 0));
-                    var badValue = (mask[ii][jj] & Math.pow(2, 5));
-                    var mathError = (mask[ii][jj] & Math.pow(2, 6));
-                    var badFit = (mask[ii][jj] & Math.pow(2, 7));
-                    var doNotUse = (mask[ii][jj] & Math.pow(2, 30));
-                    var noData = noValue;
-                    var badData = (badValue || mathError || badFit || doNotUse);
+                    var bitmasks = this.plotparams["bitmasks"];
+                    var noData = (mask[ii][jj] & Math.pow(2, bitmasks["nocov"]));
+                    var badData = false;
+                    for (var key in bitmasks["badData"]) {
+                        badData = badData || (mask[ii][jj] & Math.pow(2, bitmasks["badData"][key]))
+                    };
                 } else {
                     noData == null;
                     badData == null;
                 }
-
+                var signalToNoiseThreshold = this.plotparams["snr_min"];
                 if (ivar !== null) {
                     var signalToNoise = Math.abs(val) * Math.sqrt(ivar[ii][jj]);
-                    if (this.title.toLowerCase().indexOf("vel") >= 0) {
-                        var signalToNoiseThreshold = 0.;
-                    } else {
-                        var signalToNoiseThreshold = 1.;
-                    };
                 }
 
                 // value types
                 // val=no-data => gray color
                 // val=null => hatch area
-                // val=low-sn => nothing at the moment
 
                 if (noData) {
-                    // for data that is outside the range "nocov" mask
+                    // for data that is outside the range "NOCOV" mask
                     val = 'no-data';
                 } else if (badData) {
                     // for data that is bad - masked in some way
                     val = null;
                 } else if (ivar !== null && (signalToNoise < signalToNoiseThreshold)) {
                     // for data that is low S/N
-                   val = null ; //val = 'low-sn';
+                    val = null ;
                 } else if (ivar === null) {
                     // for data with no mask or no inverse variance extensions
                     if (this.title.search('binid') !== -1) {
@@ -251,11 +245,11 @@ class HeatMap {
         '#810823', '#7f0823', '#7c0722', '#790622', '#760521', '#730421', '#700320', '#6d0220',
         '#6a011f', '#67001f'];
 
-        if (cmap === "linearLab") {
+        if (cmap === "linearlab") {
             return linearLabHex;
         } else if (cmap === "inferno") {
             return infernoHex;
-        } else if (cmap === "RdBu") {
+        } else if (cmap === "RdBu_r") {
             return RdBuHex;
         } else {
             return ["#000000", "#FFFFFF"];
@@ -274,14 +268,8 @@ class HeatMap {
 
     quantileClip(range){
         var quantLow, quantHigh, zQuantLow, zQuantHigh;
+        [quantLow, quantHigh] = this.plotparams["percentile_clip"];
         [zQuantLow, zQuantHigh] = this.getMinMax(range);
-        if (this.title.toLowerCase().indexOf("vel") >= 0 ||
-            this.title.toLowerCase().indexOf("sigma") >= 0) {
-            [quantLow, quantHigh] = [10, 90];
-        } else if (this.title.toLowerCase().indexOf("flux") >= 0) {
-            [quantLow, quantHigh] = [5, 95];
-        };
-
         if (range.length > 0) {
             if (quantLow > 0) {
                 zQuantLow = math.quantileSeq(range, quantLow / 100);
@@ -315,16 +303,12 @@ class HeatMap {
         // [zmin, zmax] = this.getMinMax(zrange);
         [zmin, zmax] = this.quantileClip(zrange);
 
+        var cmap = this.plotparams["cmap"];
 
-        if (this.title.toLowerCase().indexOf("vel") >= 0) {
-            var cmap = "RdBu";
-            // make velocity maps symmetric
+        // make color bar symmetric
+        if (this.plotparams["symmetric"]){
             var zabsmax = Math.max.apply(null, [Math.abs(zmin), Math.abs(zmax)]);
             [zmin, zmax] = [-zabsmax, zabsmax];
-        } else if (this.title.toLowerCase().indexOf("sigma") >= 0) {
-            var cmap = "inferno";
-        } else {
-            var cmap = "linearLab";
         };
 
         var cstops = this.setColorStops(cmap);
@@ -340,7 +324,10 @@ class HeatMap {
                 plotBackgroundColor: '#A8A8A8'
             },
             credits: {enabled: false},
-            title: {text: this.title},
+            title: {
+                text: this.title.replace(/[_]/g, " "),
+                style: {fontSize: "14px"}
+            },
             navigation: {
                 buttonOptions: {
                     theme: {fill: null}